+++ packages/pwa-kit-react-sdk/src/ssr/universal/extensibility/utils.ts
@@ -8,6 +8,7 @@
 import React from 'react'
 import {applyHOCs} from '../utils'
 import ApplicationExtension from './application-extension'
+import {ApplicationExtensionConfig} from './types'
 
 /**
  * Given the provided Application, apply all the App extensions to it.
@@ -16,12 +17,9 @@ import ApplicationExtension from './application-extension'
  */
 export const applyAppExtensions = (
     App: React.ComponentType,
-    extensions: ApplicationExtension[]
+    extensions: ApplicationExtension<ApplicationExtensionConfig>[]
 ): React.ComponentType => {
     const extendAppHocs = extensions
-        // TODO: All Application Extensions configuration objects will extend from a single IApplicationExtensionConfig type so that we
-        // can add a feature to toggle if the extension is enabled or disabled. This will happen in the tupal config support ticket.
-        // .map((extension) => extension.getConfig().enabled && extension.extendApp)
         .map((extension) => extension.extendApp.bind(extension))
         .filter(Boolean)